# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TruckGo API - Uber-for-trucks platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /truck-types/{id}:
    get:
      operationId: getTruckType
      tags: [truck-types]
      summary: Get a truck type by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Truck type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TruckType"
        "404":
          description: Not found

  # --- Price Estimates ---
...
  /drivers/available:
    get:
      operationId: listAvailableDrivers
      tags: [drivers]
      summary: List nearby available drivers
      parameters:
        - in: query
          name: truckTypeId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Available drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
  /drivers/{id}/status:
    patch:
      operationId: updateDriverStatus
      tags: [drivers]
      summary: Update driver availability status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}/jobs:
    get:
      operationId: listDriverJobs
      tags: [drivers]
      summary: List all jobs (bookings) for a driver
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver's jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /drivers/{id}/stats:
    get:
      operationId: getDriverStats
      tags: [drivers]
      summary: Get driver performance stats (earnings, trips, rating)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverStats"
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # --- Admin ---
...
  /reviews/driver/{driverId}:
    get:
      operationId: listDriverReviews
      tags: [reviews]
      summary: List reviews for a driver
      parameters:
        - in: path
          name: driverId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
components:
  schemas:
...
    TruckType:
      type: object
      required: [id, name, description, capacityKg, basePrice, pricePerKm]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        capacityKg:
          type: number
        basePrice:
          type: number
        pricePerKm:
          type: number
        imageUrl:
          type: ["string", "null"]

    # --- Estimates ---
...
    Booking:
      type: object
      required: [id, customerId, truckTypeId, pickupAddress, dropoffAddress, distanceKm, estimatedPrice, status, goodsDescription, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        driverId:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
        driverPhone:
          type: ["string", "null"]
        truckTypeId:
          type: integer
        truckTypeName:
          type: ["string", "null"]
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        distanceKm:
          type: number
        estimatedPrice:
          type: number
        finalPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        goodsDescription:
          type: string
        notes:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Driver:
      type: object
      required: [id, name, phone, licenseNumber, vehiclePlate, truckTypeId, status, rating, totalTrips]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        licenseNumber:
          type: string
        vehiclePlate:
          type: string
        vehicleYear:
          type: ["integer", "null"]
        truckTypeId:
          type: integer
        truckTypeName:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, busy, offline]
        rating:
          type: number
        totalTrips:
          type: integer
        createdAt:
          type: string
          format: date-time
    DriverStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [available, busy, offline]
    DriverStats:
      type: object
      required: [driverId, totalTrips, completedTrips, cancelledTrips, totalEarnings, rating, reviewCount]
      properties:
        driverId:
          type: integer
        totalTrips:
          type: integer
        completedTrips:
          type: integer
        cancelledTrips:
          type: integer
        totalEarnings:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer

    # --- Reviews ---
    ReviewInput:
      type: object
      required: [bookingId, customerId, driverId, rating]
      properties:
        bookingId:
          type: integer
        customerId:
          type: integer
        driverId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
    Review:
      type: object
      required: [id, bookingId, customerId, driverId, rating, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        driverId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # --- Admin ---
...