# lib/api-zod/src/generated/types/truckType.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TruckGo API - Uber-for-trucks platform
 * OpenAPI spec version: 0.1.0
 */

export interface TruckType {
  id: number;
  name: string;
  description: string;
  capacityKg: number;
  basePrice: number;
  pricePerKm: number;
  /** @nullable */
  imageUrl?: string | null;
}